Handle names for enums, use longs for integer types

+++ Replicate/Serialization/JSONGraphSerializer.cs
@@ -15,8 +15,8 @@ namespace Replicate.Serialization
         public class JSONIntSerializer : ITypedSerializer
         {
             Regex rx = new Regex(@"[0-9\.\+\-eE]");
-            public object Read(Stream stream) => int.Parse(stream.ReadAllString(c => rx.IsMatch("" + c)));
-            public void Write(object obj, Stream stream) => stream.WriteString(Convert.ChangeType(obj, typeof(int)).ToString());
+            public object Read(Stream stream) => long.Parse(stream.ReadAllString(c => rx.IsMatch("" + c)));
+            public void Write(object obj, Stream stream) => stream.WriteString(Convert.ChangeType(obj, typeof(long)).ToString());
         }
         public class JSONBoolSerializer : ITypedSerializer
         {
